Fix virtual_machines attribute access in VM conflict check
Added default([]) filter and is defined check to prevent attribute
error when vmware_vm_info fails.

Error:
'Error while evaluating conditional: object of type dict has no
attribute virtual_machines'

The 'when' condition is evaluated after the 'that' clause is templated,
so we need to provide a default value and check if the attribute exists.